Add patch for subtractive synthesis
This commit will add a subctrative synthesis patch. This patch allows
the selection of some types of filter (high pass, low pass and bandpass)
and shows the output for a variety of inputs (sine wave, sawtooth wave
and white noise). It also allows to control filter parameters, like
cuttoff frequency and the Q factor (for bandpass filter).